bot/requirements.txt
discord.py==2.3.2
aiosqlite==0.20.0
